break out user models into own file
