+++ app/models/__init__.py
@@ -1,29 +1,3 @@
-from flask_login import UserMixin
-from .db import db
-from .company import CompanyMembers # Needs this for User to be able to have relationship with company_members
+
 # TODO Should we import all models into __init__.py for simplicity of importing models in the rest of the app?
 
-
-class User(UserMixin, db.Model):
-    __tablename__ = "user"
-    id = db.Column(db.Integer, primary_key=True)
-    email = db.Column(db.String(100), unique=True)
-    password = db.Column(db.String(100))
-    name = db.Column(db.String(1000))
-    roles = db.relationship('Role', secondary='user_roles')
-    companies = db.relationship('Company', secondary='company_members')
-
-
-class Role(db.Model):
-    __tablename__ = 'roles'
-    id = db.Column(db.Integer(), primary_key=True)
-    name = db.Column(db.String(50), unique=True)
-
-
-class UserRoles(db.Model):
-    __tablename__ = 'user_roles'
-    id = db.Column(db.Integer(), primary_key=True)
-    user_id = db.Column(db.Integer(), db.ForeignKey(
-        'user.id', ondelete='CASCADE'))
-    role_id = db.Column(db.Integer(), db.ForeignKey(
-        'roles.id', ondelete='CASCADE'))
